+++ docs/changelog.md
@@ -2,11 +2,13 @@
 # v7.1.1.3 Beta (????-??-??)
 
 - Support for autocreate-playlist, video-exts, audio-exts, image-exts.
-  Windows 7 support should still work, but needs auto-load-folder to be enabled.
+  Windows 7 support should still work, but needs auto-load-folder to be enabled
+  or autoload.lua.
 - The command palette user script is installable from the context menu under
   `Settings > Setup > Install Command Palette`. The command palette features
   are shown in the menu under 'View > Command Palette'.
-
+- New command to select the stream quality `Video > Stream Quality (Alt+q)`,
+  this calls the Stream Quality feature of the command palette.
 
 # v7.1.1.2 Beta (2024-10-10)